commit at 11 th june

+++ 1.intro-var-operators-flowControl-function-loops/11-typeCastingIntro.cpp
@@ -3,7 +3,7 @@ using namespace std;
 int main(){
 	// cout<<"check";
 	int a=2;
-	float b=3.5;
+	double b=3.5;
 	int c=b;
 	cout<<c<<endl;
 	bool d=b;
@@ -12,4 +12,5 @@ int main(){
 	cout<<int(b)/a;
 	
 	return 0;
+	
 }
